22-01-21

+++ LABs/lab3/text_recognizer/lit_models/base.py
@@ -100,7 +100,7 @@ class BaseLitModel(pl.LightningModule):
     def training_step(self, batch, batch_idx):
         x, y = batch
         logits = self(x) # forward 매소드 호출
-        loss = self.loss_fn(logits, y)
+        loss = self.loss_fn(logits, y.to(dtype=torch.long))
         self.log("train_loss", loss)
         self.train_acc(logits, y)
         self.log("train_acc", self.train_acc, on_step=False, on_epoch=True, prog_bar=True)
@@ -110,7 +110,7 @@ class BaseLitModel(pl.LightningModule):
     def validation_step(self, batch, batch_idx):
         x, y = batch
         logits = self(x)
-        loss = self.loss_fn(logits, y)
+        loss = self.loss_fn(logits, y.to(dtype=torch.long))
         self.log("val_loss", loss, prog_bar=True)
         self.val_acc(logits, y)
         self.log("val_acc", self.val_acc, on_step=False, on_epoch=True, prog_bar=True)
